+++ packages/api-tencent/src/index.ts
@@ -4,38 +4,9 @@ import { Translator } from '@arvinxu/translator';
 import { decodeHTMLEntities } from './utils';
 import { langMap } from './langMap';
 import { sdkTranslate } from './sdk';
+import type { TencentConfig, TencentResponseResult } from './type';
 
-export interface TencentConfig {
-  secretId?: string;
-  secretKey?: string;
-  region?: string;
-  qtk?: string;
-  qtv?: string;
-}
-
-export interface TencentResponse {
-  sessionUuid: string;
-  translate: {
-    errCode: number;
-    errMsg: string;
-    sessionUuid: string;
-    source: string;
-    target: string;
-    records: TencentRawResult;
-    full: boolean;
-    options: any;
-  };
-  dict: null;
-  suggest: null;
-  errCode: number;
-  errMsg: string;
-}
-
-export type TencentRawResult = {
-  sourceText: string;
-  targetText: string;
-  traceId: string;
-}[];
+export { TencentConfig } from './type';
 
 export class Tencent extends Translator<TencentConfig> {
   /** Translator lang to custom lang */
@@ -46,64 +17,16 @@ export class Tencent extends Translator<TencentConfig> {
     langMap.map(([translatorLang, lang]) => [lang, translatorLang]),
   );
 
-  private token: {
-    qtv: string;
-    qtk: string;
-  } = {
-    qtk: '',
-    qtv: '',
-  };
+  /**
+   * tencent 获取 token 的地址
+   * @private
+   */
+  private tokenUrl: string;
 
-  private async requestWithToken(
-    text: string,
-    from: LanguageCode,
-    to: LanguageCode,
-    config: TencentConfig,
-  ): Promise<TranslateQueryResult> {
-    this.token.qtk = config.qtk;
-    this.token.qtv = config.qtv;
+  private qtk: string;
+  private qtv: string;
 
-    const data = {
-      sourceText: text,
-      source: from,
-      target: to,
-      qtv: this.token.qtv,
-      qtk: this.token.qtk,
-    };
-    const res = await this.request.post<TencentResponse>(
-      'https://fanyi.qq.com/api/translate',
-      {
-        data,
-        requestType: 'form',
-        responseType: 'json',
-        headers: {
-          Origin: 'https://fanyi.qq.com',
-          'User-Agent':
-            'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.150 Safari/537.36',
-        },
-      },
-    );
-
-    const { errMsg, errCode, translate } = res;
-    if (errCode !== 0) {
-      return this.getErrorResult(errCode, errMsg);
-    }
-
-    const { source, records, target } = translate;
-
-    console.log(translate);
-    return {
-      success: true,
-      from: Tencent.langMapReverse.get(source) || from,
-      to: Tencent.langMapReverse.get(target) || to,
-      origin: {
-        paragraphs: records.map((item) => item.sourceText),
-      },
-      trans: {
-        paragraphs: records.map((item) => item.targetText),
-      },
-    };
-  }
+  private updateAt: number;
 
   /**
    * 使用 SDK 请求
@@ -118,7 +41,7 @@ export class Tencent extends Translator<TencentConfig> {
     from: LanguageCode,
     to: LanguageCode,
     config: TencentConfig,
-  ): Promise<TranslateQueryResult> {
+  ) {
     const data = await sdkTranslate(
       text,
       Tencent.langMap.get(from),
@@ -148,11 +71,120 @@ export class Tencent extends Translator<TencentConfig> {
     text: string,
     from: LanguageCode,
     to: LanguageCode,
-    config: TencentConfig,
+    config?: TencentConfig,
   ): Promise<TranslateQueryResult> {
-    const data = await this.requestWithSDK(text, from, to, config);
+    if (config && config.withSDK) {
+      return await this.requestWithSDK(text, from, to, config);
+    }
 
-    return data;
+    return await this.requestWithoutSDK(text, from, to);
+  }
+
+  /**
+   * 不使用 SDK 请求
+   * @param text
+   * @param from
+   * @param to
+   * @private
+   */
+  private async requestWithoutSDK(
+    text: string,
+    from: LanguageCode,
+    to: LanguageCode,
+  ): Promise<TranslateQueryResult> {
+    await this.fetchToken();
+
+    const { translate } = await this.request.post<TencentResponseResult>(
+      'https://fanyi.qq.com/api/translate',
+      {
+        requestType: 'form',
+        data: {
+          source: Tencent.langMap.get(from),
+          target: Tencent.langMap.get(to),
+          sourceText: text,
+          qtv: this.qtv,
+          qtk: this.qtk,
+        },
+        headers: {
+          'User-Agent':
+            'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.150 Safari/537.36',
+          Origin: 'https://fanyi.qq.com',
+          Connection: 'keep-alive',
+        },
+      },
+    );
+
+    const formatResult = (raw: string[]) => {
+      // 腾讯翻译器的结果结构如下
+      //   [
+      //   {
+      //     sourceText: '你好',
+      //     targetText: 'Hello. ',
+      //   },
+      //     {
+      //       sourceText: '\n',
+      //       targetText: '\n ',
+      //     },
+      //     {
+      //       sourceText: '我爱你',
+      //       targetText: 'I love you',
+      //     },
+      //   ];
+      // 因此采用合并然后按 \n 切分的思路处理段落问题
+
+      return raw
+        .join('')
+        .split('\n')
+        .map((i) => i.trim());
+    };
+
+    return {
+      success: true,
+      from: Tencent.langMapReverse.get(translate.source) || from,
+      to: Tencent.langMapReverse.get(translate.target) || to,
+      text,
+      trans: {
+        paragraphs: formatResult(translate.records.map((r) => r.targetText)),
+      },
+      origin: {
+        paragraphs: formatResult(translate.records.map((r) => r.sourceText)),
+      },
+    };
+  }
+
+  async fetchToken() {
+    if (!this.tokenUrl) {
+      await this.fetchTokenUrl();
+    }
+
+    const data = await this.request
+      .post<{ qtk: string; qtv: string }>(
+        `https://fanyi.qq.com/api/${this.tokenUrl}`,
+      )
+      .catch((e) => {
+        console.error(e);
+        return { qtv: '', qtk: '' };
+      });
+
+    const { qtk, qtv } = data;
+    this.qtk = qtk;
+    this.qtv = qtv;
+    this.updateAt = Date.now();
+
+    return { qtk, qtv };
+  }
+  /**
+   * 获取请求 token 的 url
+   */
+  async fetchTokenUrl() {
+    const data = await this.request('https://fanyi.qq.com/');
+
+    const reauthuri = /reauthuri\s=\s"(.*)"/.exec(data);
+
+    // eslint-disable-next-line prefer-destructuring
+    this.tokenUrl = reauthuri[1];
+
+    return reauthuri[1];
   }
 
   readonly name = 'tencent';