+++ packages/api-tencent/CHANGELOG.md
@@ -1,8 +1,40 @@
 # 更新日志
 
-# @arvinxu/translator-tencent 1.0.0 (2021-02-17)
+## @arvinxu/translator-tencent [1.1.2](https://github.com/arvinxx/translator/compare/@arvinxu/translator-tencent@1.1.1...@arvinxu/translator-tencent@1.1.2) (2021-03-25)
 
+### ✅ 测试
+
+- 修正测试 timeout 时间 ([6676357](https://github.com/arvinxx/translator/commit/6676357))
+- 修正测试 timeout 时间 ([97ee25e](https://github.com/arvinxx/translator/commit/97ee25e))
+
+### 🐛 修复
+
+- 修复腾讯翻译器段落翻译有问题的 bug ([eee1e5a](https://github.com/arvinxx/translator/commit/eee1e5a))
+
+## @arvinxu/translator-tencent [1.1.1](https://github.com/arvinxx/translator/compare/@arvinxu/translator-tencent@1.1.0...@arvinxu/translator-tencent@1.1.1) (2021-03-24)
+
+### 🐛 修复
+
+- 修复腾讯引擎不出文字的 bug ([671b30f](https://github.com/arvinxx/translator/commit/671b30f))
+
+# @arvinxu/translator-tencent [1.1.0](https://github.com/arvinxx/translator/compare/@arvinxu/translator-tencent@1.0.1...@arvinxu/translator-tencent@1.1.0) (2021-03-22)
 
 ### ✨ 新特性
 
-* 添加腾讯 API ([a967a7e](https://github.com/arvinxx/translator/commit/a967a7e))
+- 补充腾讯 API ([9c31064](https://github.com/arvinxx/translator/commit/9c31064))
+
+## @arvinxu/translator-tencent [1.0.1](https://github.com/arvinxx/translator/compare/@arvinxu/translator-tencent@1.0.0...@arvinxu/translator-tencent@1.0.1) (2021-02-17)
+
+### 🐛 修复
+
+- 修正模块发布时是空包的 bug ([cfb1f98](https://github.com/arvinxx/translator/commit/cfb1f98))
+
+### Dependencies
+
+- **@arvinxu/translator:** upgraded to 1.1.1
+
+# @arvinxu/translator-tencent 1.0.0 (2021-02-17)
+
+### ✨ 新特性
+
+- 添加腾讯 API ([a967a7e](https://github.com/arvinxx/translator/commit/a967a7e))